util function for parsing selected components from jetty server response
